fix(crypto): force-encrypt all secret fields when --aes-key is provided
Even if you passed --aes-key with an empty database, the `host` and `default` Host keys where
never encrypted.

Now :
1) Automatically generated `host` and `default` Host key will now be encrypted by SSHportal (if you passed --aes-key)
2) At any point if you pass --aes-key SSHportal will encrypt all Password / PrivKey which were not previously encrypted

WARNING: we can't really test if we are decrypting the CFB-encrypted Password field with
the right AES key (without trying to connect to the host) so if you only have CFB-encrypted Passwords in your SSHportal DB
AND provide a bad --aes-key then your passwords will be lost! If you have at least one encrypted SSHKey the mismatch key will
be detected.

Signed-off-by: libvoid <135131094+libvoid@users.noreply.github.com>