fix(crypto): add more robusts checks to identify non-encrypted passwords
The original CFB decryption function was bad because we can't really detect if a Password field has been encrypted with a unauthenticated cipher so we have to rely on checking if the field is base64 encoded. The problem is that a plaintext base64-encoded password
can be mis-detected as CFB encrypted and SSHportal will re-encrypt a borked field.

2 new checks are now added:
1) Short base64-compatible passwords will be detected by verifying if the size of the ciphertext is lower than aes.BlockSize (16 bytes)
2) All decrypted fields should be valid UTF8 strings