chg(crypto): `DecryptField()` should always decrypt except if --aes-key is empty
The old DecryptField function returned without error the field if it believes it's unencrypted (by using a basic base64 check)

It was needed before because the SSHportal DB could contain encrypted and unencrypted secret fields and SShportal used a non-authenticated cipher.

Since the AES-GCM re-encryption this is no longer needed as AES-GCM is AEAD and all secrets are now encrypted if you set an --aes-key.
This is also a weird behavior. a decryption function should always fail if the decryption does not work.

In the case you use the server without an AES and then begin to a use one, the MigrateToGCMCipher() function will automatically
encrypt all your secrets on SSHportal startup.