fix(crypto): workaround of a bug on tx.RowsAffected in GORM
This check was initially added to detect if a row we found earlier in the function has been deleted in the meantime.
GORM update query doesn't return an error if the row you want to update doesn't exist anymore.

If you are certain your query will update the row, you can check if tx.RowsAffected == 0 to detect the above issue

Unfortunately there is a very old unresolved bug that causes tx.RowsAffected to sometimes return 0, except it's false...

I encountered this bug with a SSHportal test DB

https://github.com/go-gorm/gorm/issues/2038